fix:changed rich editor to tiptap
